% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.r
\name{build}
\alias{build}
\title{Build a model from the contents of a workbook}
\usage{
build(workbook)
}
\arguments{
\item{workbook}{File path of the workbook. The file type is guessed from
the extension which must be '.xlsx' or '.ods'. See below for the mandatory
worksheets that must be present in the workbook.}
}
\value{
An object of class \code{\link[rodeo]{rodeo}}.
}
\description{
The function builds a \code{\link[rodeo]{rodeo}}-based model by importing
all declarations and equations from a workbook established with common
spreadsheet software.
}
\note{
The file provided as \code{workbook} must contain at least the four
  mandatory sheets:
\itemize{
  \item{'vars'} Defines the state variables of the model. Mandatory columns
     are 'name', 'unit', 'description', 'default'.
  \item{'pars'} Defines the parameters of the model. Mandatory columns
     are the same as for sheet 'vars'.
  \item{'funs'} Declares and possibly defines functions used in the model
     equations. Mandatory columns are 'name' and 'code'.
  \item{'eqns'} Specifies the model equations. Mandatory columns
     are 'name', 'unit', 'description', 'rate' plus one column for
     every state variable of the model. The 'rate' columns holds the
     process rate expressions and columns named after state variables
     contain the corresponding stoichiometric factors.
}
The best way to understand the contents of the workbook is to study the
examples in the folder 'models' shipped with the package. Type
\code{system.file("models", package="rodeoEasy")} at the R prompt to see
where this folder is installed on your system.

The sheet 'funs' may contain a mixture of both intrinsic and user defined
functions. Intrinsic functions can be declared with a single line and the
column 'code' can be filled with an R comment like, e.g. '# intrinsic',
to show that the existing function will not be overloaded.
}
\examples{

x <- build(
  system.file("models/oxygen.xlsx", package="rodeoEasy")
)
}
\seealso{
Look at \code{\link{run.scenarios}} and other functions of this
  package for how to run a simulation using the built model.
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}
