% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildFromWorkbook.r
\name{buildFromWorkbook}
\alias{buildFromWorkbook}
\title{Build model from workbook}
\usage{
buildFromWorkbook(workbook)
}
\arguments{
\item{workbook}{File path of the worksheet holding the model. This can be
a file with either extension '.xlsx' or '.ods'. See below for the
mandatory file contents.}
}
\value{
An object of class \code{\link[rodeo]{rodeo}}.
}
\description{
The function builds a rodeo-based model from a workbook without
triggering a simulation.
}
\note{
The file provided as \code{workbook} must contain at least the four
  mandatory sheets:
\itemize{
  \item{'vars'} Defines the state variables of the model. Mandatory columns
     are 'name', 'unit', 'description'.
  \item{'pars'} Defines the parameters of the model. Mandatory columns
     are the same as for sheet 'vars'.
  \item{'funs'} Declares the names of functions used in the model
     equations. Mandatory columns are the same as for sheet 'vars'.
  \item{'eqns'} Specifies the model equations. Mandatory columns
     are 'name', 'unit', 'description', 'rate' plus one column for
     every state variable of the model. The 'rate' columns holds the
     process rate expressions and columns named after state variables
     contain the corresponding stoichiometric factors.
}
The best way to understand the contents of the workbook is to study the
examples in the folder 'models' shipped with the package. Type
\code{system.file("models", package="rodeoEasy")} at the R prompt to see
where the folder 'models' is installed on your system.
}
\examples{

x <- buildFromWorkbook(
  system.file("models/oxygen.xlsx", package="rodeoEasy")
)
}
\seealso{
Look at \code{\link{compareScenarios}} for building and running
  a model with a single call.
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}
