% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareScenarios.r
\name{compareScenarios}
\alias{compareScenarios}
\title{Compare scenarios}
\usage{
compareScenarios(
  workbook,
  scenarios,
  times,
  plot.vars = FALSE,
  plot.pros = FALSE,
  leg = "topright"
)
}
\arguments{
\item{workbook}{File path of the worksheet holding the model. See the
\code{\link{buildFromWorkbook}} for details on the expected file
contents. See below for further details.}

\item{scenarios}{Character vector with each element defining the name
of a 'scenario'. See details below.}

\item{times}{Numeric vector defining the times for which the future
state of the system is computed.}

\item{plot.vars}{Logical. Plot the dynamics of all state variables?}

\item{plot.pros}{Logical. Plot the dynamics of process rates?}

\item{leg}{Keyword to set the position of the legend (if plots are created).}
}
\value{
A data frame with at least three columns and one row for each
  time requested via the \code{times} argument. The first column
  indicates the scenario, the second column holds the time, and further
  columns hold the values of state variables and process rates.
}
\description{
The function builds a rodeo-based model and runs it for one
or multiple scenarios. Individual scenarios can differ by the
initial state or the values of parameters.
}
\note{
For each scenario listed in the vector \code{scenarios}, the sheet
   'vars' in \code{workbook} must contain a designated column. The name
   of that column must be 'value.' followed by the scenario identifier,
   for example 'value.default' if the scenario was titled 'default'.
   The same applies to the sheet 'pars' in \code{workbook}.
}
\examples{

x <- compareScenarios(
  system.file("models/oxygen.xlsx", package="rodeoEasy"),
  scenarios=c("default","alternative"),
  times=seq(0, 12, 0.1),
  plot.vars=TRUE
)
}
\seealso{
Look at \code{\link{buildFromWorkbook}} for building a model
  without running it immediately.
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}
