% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_package.r
\docType{package}
\name{rodeoEasy-package}
\alias{rodeoEasy-package}
\title{ODE-based models in teaching}
\description{
The package provides high-level functions to build and run
models rooted in simultaneous ODE. It is primarily meant for
teaching purposes.
}
\author{
\email{david.kneis@tu-dresden.de}
}
