% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.scenarios.r
\name{show.scenarios}
\alias{show.scenarios}
\title{Visualize scenarios for a single variable}
\usage{
show.scenarios(
  x,
  variable,
  xlab = "time",
  ylab = variable,
  ylog = FALSE,
  ylim = NULL
)
}
\arguments{
\item{x}{A data frame returned by \code{\link{run.scenarios}}.}

\item{variable}{The name of a single variable (character string).
A column with that name must be present in \code{x}.}

\item{xlab}{Label for the x axis (character string).}

\item{ylab}{Label for the y axis (character string).}

\item{ylog}{Logical. If \code{TRUE}, the y axis will be log-scaled.}

\item{ylim}{Either a numeric vector of length two used to trim the y axis
or \code{NULL} to adjust the range automatically.}
}
\value{
A vector of colors with element names corresponding to the
  names of scenarios. This is useful for building a legend.
}
\description{
The function takes a data frame returned by \code{\link{run.scenarios}}
and plots the dynamics of a single variable. The different scenarios
are distinguished by color.
}
\examples{

# build the model
m <- build(system.file("models/oxygen.xlsx", package="rodeoEasy"))

# implement specific functions used in RHS of ODE
DO_sat <- function(T) { exp(7.712 - 1.314 * log(T + 45.93)) }

# run scenarios WITHOUT plotting
x <- run.scenarios(model=m, times=seq(0, 12, 0.1),
  scenarios=list(warm=c(kd=1.25), cold.windy=c(kd=1, ka=5)),
  plot.vars=FALSE
)

# custom plotting
par(mfrow=c(1,2))
show.scenarios(x, "DO")
keyColors <- show.scenarios(x, "OM")
legend("topright", bty="n", lty=1, col=keyColors,
  legend=names(keyColors))
par(mfrow=c(1,1))
}
\seealso{
Look at \code{\link{run.scenarios}} to create a suitable
  input data frame.
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}
